Gap 15: nuint addressing past 2 GiB on NativePointer memories
MemoryHelpers.StoreI32 / LoadI32 (and every load/store/narrow/F32/F64
sibling) cast the effective address to int on the final
mem.RefAs<byte>(...) / mem.AsSpan(...) call. With ea > int.MaxValue
— anything past 2 GiB into a NativePointer-backed linear memory —
that cast wrapped to a negative pointer offset; the kernel signaled
SIGSEGV and the .NET runtime aborted with AccessViolationException.
Bypassed managed exception handling, so the wasm-trap-to-exit-1
path didn't catch it.

The bounds check itself was already correct — ea is computed as
long via (uint)addr + offset, and BoundsCheck compares against
mem.ByteLength (nuint, mode-aware). Only the truncating (int)ea
cast on the access call site was wrong.

MemoryInstance gains nuint sibling overloads:
  - RefAs<T>(nuint ea): byte* + nuint pointer arithmetic on
    NativeBase; ManagedArray branch keeps the safe (int)ea cast
    (Array.MaxLength bounds the byte[] backing ≤ 2 GiB).
  - AsSpan(nuint offset, int length): same shape for narrow
    load/store siblings (StoreI32_8, etc.).

Migrated call sites (all switch to (nuint)ea):
  - Wacs.Transpiler.Lib/AOT/Emitters/MemoryEmitter.cs
    MemoryHelpers — 59 sites across i32/i64 + every narrow variant
    + f32/f64
  - Wacs.Core/Wacs.Core/Instructions/Memory/{I32,I64,F}MemoryLoad.cs
    + Inst{I32,I64}Store.cs + FMemoryStore.cs — interpreter
    per-instruction handlers had the same shape; ~30 sites

Test surface: new
Wacs.Transpiler.Test.MemoryHelpersHighAddressTests covers
StoreI32/LoadI32, StoreI64/LoadI64, StoreI32_8/LoadI32_8U,
StoreF32/LoadF32, StoreF64/LoadF64 at ea = 0x80000400 (~2 GiB +
1 KiB) on a NativePointer 33000-page (~2.0625 GiB) instance.
Pre-fix every test AVs; post-fix all five round-trip.
NativeMemory.AllocZeroed is lazy-zero on calloc so the 2 GiB
virtual reservation does not commit physical pages.

Versions:
  - WACS                0.13.2 → 0.13.3
  - WACS.Cli            1.5.2  → 1.5.3
  - WACS.Transpiler.Lib 0.8.2  → 0.8.3

Test counts: Wacs.Core 394, Wacs.ComponentModel 354,
Wacs.Transpiler 764 (+5), Wacs.WASI.Preview2 189, Wacs.HostBindings
14, Spec.Test 770. All green. AOT smoke (wacs aot --wasi) still
produces a working native binary (5.8 MB osx-arm64, "hello"
output, exit 0).

Out of scope (separate gap): atomics. AtomicHelpers.CheckEa
returns int and the int ea cascades through the abstract
InstAtomicLoad.DoLoad / InstAtomicStore.DoStore signatures. Same
shape as gap 15 but a different cohort of guests (atomic-using
shared-memory threading); follow-up.

+++ Wacs.Core/Wacs.Core/Instructions/Memory/FMemoryLoad.cs
@@ -50,7 +50,7 @@ namespace Wacs.Core.Instructions.Memory
             long ea = offset + M.Offset;
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer {ea}+{WidthTByteSize} out of bounds ({(long)mem.ByteLength}).");
-            var bs = mem.AsSpan((int)ea, WidthTByteSize);
+            var bs = mem.AsSpan((nuint)ea, WidthTByteSize);
             
             return MemoryMarshal.Read<float>(bs);
         }
@@ -84,7 +84,7 @@ namespace Wacs.Core.Instructions.Memory
             long ea = offset + M.Offset;
             if ((ulong)ea > (ulong)mem.ByteLength || (ulong)mem.ByteLength - (ulong)ea < (ulong)WidthTByteSize)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer {ea}+{WidthTByteSize} out of bounds ({(long)mem.ByteLength}).");
-            var bs = mem.AsSpan((int)ea, WidthTByteSize);
+            var bs = mem.AsSpan((nuint)ea, WidthTByteSize);
             
             return MemoryMarshal.Read<double>(bs);
         }
